+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - added add/cut/intersect_sketch()
 - added make_sphere() and make_cylinder()
 - added cut_hole()
+- added Body.move_to_origin() and Body.center()
 
 ## [0.4] - 2024-01-12